+++ site/dat/docs/Changelog.md
@@ -6,6 +6,7 @@
  - fix public cloud reporting (unauthorised notes sending)
  - add test for paths of resource files
  - bump up auto-installed JMeter to 3.1 and pmgr to 0.11
+ - add limitation of response time list size to cumulative KPISets
  - fix the crash when multiple BM reporters are in use
  - add `public-report` option to BlazeMeter reporter and cloud provisioning
  - add `retrieve-resources-regex` scenario-level option to JMeter